refactor(analytics): split topix sma ratio research helpers
